requirements.txt
aiohttp
google-generativeai